+++ web-app/static/form.js
@@ -1,12 +1,9 @@
 //To stay in the same page
 
 $(document).ready(function () {
-
       
       
-      // var video = document.getElementById('video');
-      // get_face_descriptor(video);
-      // console.log(fullFaceDescriptions);
+     
 });
 // function form() {
 //     // var form_data = $('form').serializeArray().reduce(function(obj, item) {
