+++ web-app/templates/admin.html
@@ -9,6 +9,7 @@
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/socket.io-client@2/dist/socket.io.js"></script>
     <style>
         .fakeimg {
             height: 200px;
@@ -79,7 +80,6 @@
     </div>
 </footer>
 
-<script src="static/live_video.js"></script>
 <script src="static/face-api.min.js"></script>
 <script src="static/get_face_descriptor.js"></script>
 <script src="static/form.js"></script>
